Implemented alternate action space as such: Action 0: Change the prep pulse FA Action 1: Change the acquisition pulse base FA Actions 2-n: Change the acquisition pulse delta nodes
The acquisition pulses are thus now controlled via a "base" FA value
and a set of delta nodes instead of merely the nodes.
It is expected that this will enable better contrast control, as
the base FA is now controlled by a single output neuron instead
of all but one.